Several MP fixes, resolve a TODO, cleanup * Fix minions still running AI for one tick after player despawns, causing exceptions * Fix bow of friendship usage in multiplayer causing server and client exceptions * Move some code back into ModSystem as it's now fixed on tmls side * Init reference fields in LeveledCombatPetModPlayer correctly (gameplay: set starting tactic correctly) * Load certain classes only on client

+++ Projectiles/Minions/LilEnt/LilEnt.cs
@@ -154,13 +154,15 @@ namespace AmuletOfManyMinions.Projectiles.Minions.LilEnt
 			base.TargetedMovement(vectorToTargetPosition);
 		}
 
-		public override void CheckActive()
+		public override bool CheckActive()
 		{
-			base.CheckActive();
-			if(!player.GetModPlayer<MinionSpawningItemPlayer>().lilEntAccessoryEquipped)
+			if (base.CheckActive() && !player.GetModPlayer<MinionSpawningItemPlayer>().lilEntAccessoryEquipped)
 			{
 				Projectile.Kill();
+				return false;
 			}
+
+			return true;
 		}
 	}
 }
